Refactor: file name and notice message
