+++ app/(dashboard)/accounts/page.tsx
@@ -5,7 +5,7 @@ import {Card, CardContent, CardHeader, CardTitle} from "@/components/ui/card";
 import {Button} from "@/components/ui/button";
 import {useNewAccount} from "@/features/accounts/hooks/use-new-account";
 import {useGetAccounts} from "@/features/accounts/api/use-get-accounts";
-import {useBulkDeleteAccounts} from "@/features/accounts/api/use-bulk-delete";
+import {useBulkDeleteAccounts} from "@/features/accounts/api/use-bulk-delete-accounts";
 
 import {columns} from "@/app/(dashboard)/accounts/columns";
 import {DataTable} from "@/components/ui/data-table";
