+++ features/accounts/api/use-bulk-delete-accounts.ts
