+++ features/accounts/components/edit-account-sheet.tsx
@@ -28,7 +28,7 @@ export const EditAccountSheet = () => {
 
   const [ConfirmDialog, confirm] = useConfirm(
     "Are you sure?",
-    "You are about to delete this transaction."
+    "You are about to delete this account."
   )
 
   const accountQuery = useGetAccount(id)
